End of website
Sending a message works but does not save into the child attribute.
Marked as it can be an empty value and the message get passed. My only
issue is how to know what the user chose from the dropdown. I searched
on codeigniter and tried to apply code but did not manage to make it
work. I am only missing one thing .. SAVING THE RECEPIENt ID. EVerything
else works just fine even the display.